change image data

+++ render/util/coordinate_grid.cc
@@ -9,7 +9,7 @@
 namespace azer {
 
 // class CoordinateGrid
-CoordinateGrid::CoordinateGrid(int width, int height, int num)
+CoordinateGrid::CoordinateGrid(float width, float height, int num)
     : x_color_(Vector4(1.0f, 0.0f, 0.0f, 1.0)),
       z_color_(Vector4(0.0f, 0.0f, 1.0f, 1.0)),
       kWidth(width), kHeight(height), kNum(num) {
@@ -28,21 +28,21 @@ void CoordinateGrid::Init() {
   data_ = new VertexData(effect->GetVertexDesc(), (kNum + 1) * 4);
   CoordinateGridEffect::Vertex* vertex =
       (CoordinateGridEffect::Vertex*)data_->pointer();
-  float row_width = (float)kHeight / (float)kNum;
+  float row_width = kHeight / (float)kNum;
   for (int i = 0; i < kNum + 1; ++i) {
-    float z = (float)kHeight / 2.0f - row_width * i;
-    vertex->position = Vector4(-(float)kWidth / 2.0f, 0.0f, z, 1.0f);
+    float z = kHeight / 2.0f - row_width * i;
+    vertex->position = Vector4(-kWidth / 2.0f, 0.0f, z, 1.0f);
     vertex++;
-    vertex->position = Vector4((float)kWidth / 2.0f, 0.0f, z, 1.0f);
+    vertex->position = Vector4(kWidth / 2.0f, 0.0f, z, 1.0f);
     vertex++;
   }
 
-  float col_width = (float)kHeight / (float)kNum;
+  float col_width = kHeight / (float)kNum;
   for (int i = 0; i < kNum + 1; ++i) {
-    float x  = (float)kWidth / 2.0f - col_width * i;
-    vertex->position = Vector4(x, 0.0f, -(float)kHeight / 2.0f, 1.0f);
+    float x  = kWidth / 2.0f - col_width * i;
+    vertex->position = Vector4(x, 0.0f, -kHeight / 2.0f, 1.0f);
     vertex++;
-    vertex->position = Vector4(x, 0.0f, (float)kHeight / 2.0f, 1.0f);
+    vertex->position = Vector4(x, 0.0f, kHeight / 2.0f, 1.0f);
     vertex++;
   }
   vb_ = rs->CreateVertexBuffer(VertexBuffer::Options(), data_.get());
